refactor(transfer): component optimization
