+++ packages/components/time-picker/panel/single-panel.tsx
@@ -29,7 +29,6 @@ export default defineComponent({
     resetTriggerScroll: Function,
     isShowPanel: Boolean,
   },
-
   setup(props) {
     const { globalConfig } = useConfig('timePicker');
     const COMPONENT_NAME = usePrefixClass('time-picker__panel');
