+++ packages/components/timeline/hooks/index.ts
@@ -1,5 +1,5 @@
 import { InjectionKey, ComputedRef, computed } from 'vue';
-import { TdTimelineProps } from './type';
+import { TdTimelineProps } from '../type';
 
 export const DefaultAlign = {
   vertical: ['left', 'right'],
