+++ packages/components/tooltip/utils/index.ts
@@ -1,5 +1,5 @@
 import { onMounted, onUnmounted, ref } from 'vue';
-import { isServer } from '../utils/dom';
+import { isServer } from '../../utils/dom';
 
 export const useMouse = () => {
   const x = ref(0);
