+++ packages/components/transfer/hooks/useDragSort.ts
@@ -1,6 +1,6 @@
 import { ref, ComputedRef, Ref } from 'vue';
 import type { TransferValue } from '../type';
-import type { TransferItemOption } from '../interface';
+import type { TransferItemOption } from '../types';
 export default function useDragSort(
   currentValue: Ref<Array<TransferValue>>,
   curPageData: ComputedRef<TransferItemOption[]>,
