+++ packages/components/transfer/utils/index.ts
@@ -3,9 +3,9 @@ import { cloneDeep } from 'lodash-es';
 import { isUndefined } from 'lodash-es';
 import { isArray } from 'lodash-es';
 
-import { TransferListOptionBase, TransferItemOption, TdTransferProps, TransferValue, DataOption } from './interface';
+import { TransferListOptionBase, TransferItemOption, TdTransferProps, TransferValue, DataOption } from '../types';
 
-export { emitEvent } from '../utils/event';
+export { emitEvent } from '../../utils/event';
 
 export const TRANSFER_NAME = 'TTransfer';
